// libs/peering/api/src/lib/peer-importer/graphql.ts
export type Maybe<T> = T | null
export type InputMaybe<T> = Maybe<T>
export type Exact<T extends {[key: string]: unknown}> = {[K in keyof T]: T[K]}
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & {[SubKey in K]?: Maybe<T[SubKey]>}
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & {[SubKey in K]: Maybe<T[SubKey]>}
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string
  String: string
  Boolean: boolean
  Int: number
  Float: number
  /** A hexidecimal color value. */
  Color: string
  /** A date string, such as 2007-12-03, compliant with the `full-date` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  Date: string
  /** A date-time string at UTC, such as 2019-12-03T09:54:33Z, compliant with the date-time format. */
  DateTime: string
  /** Setting Value */
  GraphQLSettingValueType: any
  RichText: Node[]
  Slug: string
  /** The `Upload` scalar type represents a file upload. */
  Upload: File
  /** A valid vote value */
  VoteValue: any
  _Any: any
  _FieldSet: any
  link__Import: any
}

export type AllowedSettingVals = {
  __typename?: 'AllowedSettingVals'
  boolChoice?: Maybe<Scalars['Boolean']>
  stringChoice?: Maybe<Array<Scalars['String']>>
}

export type Article = {
  __typename?: 'Article'
  createdAt: Scalars['DateTime']
  disableComments: Scalars['Boolean']
  draft?: Maybe<ArticleRevision>
  hidden: Scalars['Boolean']
  id: Scalars['ID']
  modifiedAt: Scalars['DateTime']
  peer?: Maybe<Peer>
  peerArticleId?: Maybe<Scalars['String']>
  peerId?: Maybe<Scalars['String']>
  pending?: Maybe<ArticleRevision>
  published?: Maybe<ArticleRevision>
  shared: Scalars['Boolean']
  tags: Array<Tag>
  url: Scalars['String']
}
// ...
export type ArticleRevision = HasBlockContent & {
  __typename?: 'ArticleRevision'
  authors: Array<Author>
  blocks: Array<BlockContent>
  breaking: Scalars['Boolean']
  canonicalUrl: Scalars['String']
  createdAt: Scalars['DateTime']
  hideAuthor: Scalars['Boolean']
  id: Scalars['ID']
  image?: Maybe<Image>
  imageID?: Maybe<Scalars['String']>
  lead?: Maybe<Scalars['String']>
  preTitle?: Maybe<Scalars['String']>
  properties: Array<Property>
  publishedAt: Scalars['DateTime']
  seoTitle?: Maybe<Scalars['String']>
  slug?: Maybe<Scalars['String']>
  socialMediaAuthors: Array<Author>
  socialMediaDescription?: Maybe<Scalars['String']>
  socialMediaImage?: Maybe<Image>
  socialMediaImageID?: Maybe<Scalars['String']>
  socialMediaTitle?: Maybe<Scalars['String']>
  title?: Maybe<Scalars['String']>
}
// ...
export type Author = {
  __typename?: 'Author'
  bio?: Maybe<Scalars['RichText']>
  createdAt: Scalars['DateTime']
  hideOnArticle?: Maybe<Scalars['Boolean']>
  hideOnTeam?: Maybe<Scalars['Boolean']>
  hideOnTeaser?: Maybe<Scalars['Boolean']>
  id: Scalars['ID']
  image?: Maybe<Image>
  jobTitle?: Maybe<Scalars['String']>
  links?: Maybe<Array<AuthorLink>>
  modifiedAt: Scalars['DateTime']
  name: Scalars['String']
  slug: Scalars['Slug']
  tags: Array<Tag>
  url: Scalars['String']
}

export type AuthorConnection = {
  __typename?: 'AuthorConnection'
  nodes: Array<Author>
  pageInfo: PageInfo
  totalCount: Scalars['Int']
}

export type AuthorFilter = {
  hideOnTeam?: InputMaybe<Scalars['Boolean']>
  name?: InputMaybe<Scalars['String']>
  tagIds?: InputMaybe<Array<Scalars['ID']>>
}

export type AuthorLink = {
  __typename?: 'AuthorLink'
  title: Scalars['String']
  url: Scalars['String']
}
// ...
export type BlockContent =
  | BildwurfAdBlock
  | BreakBlock
  | CommentBlock
  | EventBlock
  | FacebookPostBlock
  | FacebookVideoBlock
  | HtmlBlock
  | IFrameBlock
  | ImageBlock
  | ImageGalleryBlock
  | InstagramPostBlock
  | ListicleBlock
  | PolisConversationBlock
  | PollBlock
  | QuoteBlock
  | RichTextBlock
  | SoundCloudTrackBlock
  | TeaserGridBlock
  | TeaserGridFlexBlock
  | TikTokVideoBlock
  | TitleBlock
  | TwitterTweetBlock
  | UnknownBlock
  | VimeoVideoBlock
  | YouTubeVideoBlock
// ...
export type FocalPoint = {
  __typename?: 'FocalPoint'
  x: Scalars['Float']
  y: Scalars['Float']
}
// ...
export type HasBlockContent = {
  blocks: Array<BlockContent>
}
// ...
export type Image = {
  __typename?: 'Image'
  createdAt: Scalars['DateTime']
  description?: Maybe<Scalars['String']>
  extension: Scalars['String']
  fileSize: Scalars['Int']
  filename?: Maybe<Scalars['String']>
  focalPoint?: Maybe<FocalPoint>
  format: Scalars['String']
  height: Scalars['Int']
  id: Scalars['ID']
  license?: Maybe<Scalars['String']>
  link?: Maybe<Scalars['String']>
  mimeType: Scalars['String']
  modifiedAt: Scalars['DateTime']
  source?: Maybe<Scalars['String']>
  tags: Array<Scalars['String']>
  title?: Maybe<Scalars['String']>
  transformURL?: Maybe<Scalars['String']>
  url?: Maybe<Scalars['String']>
  width: Scalars['Int']
}
// ...
export type Peer = {
  __typename?: 'Peer'
  createdAt: Scalars['DateTime']
  hostURL: Scalars['String']
  id: Scalars['ID']
  isDisabled?: Maybe<Scalars['Boolean']>
  modifiedAt: Scalars['DateTime']
  name: Scalars['String']
  profile?: Maybe<PeerProfile>
  slug: Scalars['String']
}
// ...
export type PeerProfile = {
  __typename?: 'PeerProfile'
  callToActionImage?: Maybe<Image>
  callToActionImageURL?: Maybe<Scalars['String']>
  callToActionText: Scalars['RichText']
  callToActionURL: Scalars['String']
  hostURL: Scalars['String']
  logo?: Maybe<Image>
  name: Scalars['String']
  themeColor: Scalars['Color']
  themeFontColor: Scalars['Color']
  websiteURL: Scalars['String']
}
// ...
export type Property = {
  __typename?: 'Property'
  id: Scalars['String']
  key: Scalars['String']
  public: Scalars['Boolean']
  value: Scalars['String']
}
// ...
export type Query = {
  __typename?: 'Query'
  _entities: Array<Maybe<_Entity>>
  _service: _Service
  /**
   *
   *       Returns all active subscribers.
   *       Includes subscribers with a cancelled but not run out subscription.
   *
   */
  activeSubscribers: Array<DashboardSubscription>
  /** Returns an article by id. */
  article: Article
  /** Returns a paginated list of articles based on the filters given. */
  articles: PaginatedArticles
  /** This query returns the redirect Uri. */
  authProviders: Array<AuthProvider>
  /** This query takes either the ID or the slug and returns the author. */
  author?: Maybe<Author>
  /** This query is to get the authors. */
  authors: AuthorConnection
// ...
  tags?: Maybe<TagConnection>
  /**
   *
   *       Returns a single userConsent by id.
   *
   */
  userConsent: UserConsent
  /**
   *
   *       Returns a list of userConsents. Possible to filter.
   *
   */
  userConsents: Array<UserConsent>
  /** This query returns the answerId of a poll if the user has already voted on it. */
  userPollVote?: Maybe<Scalars['ID']>
  versionInformation: VersionInformation
}
// ...
export type Tag = {
  __typename?: 'Tag'
  id: Scalars['ID']
  main: Scalars['Boolean']
  tag?: Maybe<Scalars['String']>
  type?: Maybe<TagType>
  url: Scalars['String']
}

export type TagConnection = {
  __typename?: 'TagConnection'
  nodes: Array<Tag>
  pageInfo: PageInfo
  totalCount: Scalars['Int']
}

export type TagFilter = {
  tag?: InputMaybe<Scalars['String']>
  type?: InputMaybe<TagType>
}

export enum TagSort {
  CreatedAt = 'CreatedAt',
  ModifiedAt = 'ModifiedAt',
  Tag = 'Tag'
}

export enum TagType {
  Article = 'Article',
  Author = 'Author',
  Comment = 'Comment',
  Event = 'Event',
  Page = 'Page'
}
// ...
export type ArticleQueryVariables = Exact<{
  id: Scalars['String']
}>

export type ArticleQuery = {
  __typename?: 'Query'
  article: {
    __typename?: 'Article'
    id: string
    tags: Array<{__typename?: 'Tag'; tag?: string | null}>
    published?: {
      __typename?: 'ArticleRevision'
      authors: Array<{
        __typename?: 'Author'
        name: string
        slug: string
        bio?: Node[] | null
        jobTitle?: string | null
        hideOnArticle?: boolean | null
        image?: {__typename?: 'Image'; url?: string | null} | null
      }>
    } | null
  }
}

export const Article = gql`
  query Article($id: String!) {
    article(id: $id) {
      id
      tags {
        tag
      }
      published {
        authors {
          name
          slug
          bio
          jobTitle
          image {
            url
          }
          hideOnArticle
        }
      }
    }
  }
`